fix: lint, move userdropdown
